enregistrement user et session

+++ db/migrations/20200918100513_create_users.sql
@@ -2,7 +2,7 @@
 -- SQL in section 'Up' is executed when this migration is applied
 CREATE TABLE users(
   id BIGSERIAL PRIMARY KEY,
-  gitlab_id INT,
+  gitlab_id BIGINT,
   name VARCHAR NOT NULL,
   email VARCHAR NOT NULL,
   avatar_url VARCHAR ,
